Item update policy in download mode / channel info update policy / channel context menu
Item update policy in download mode
-----------------------------------
To minimize race-condition, download first -> insert item to DB -> rename file based on item id.

Channel info update policy
--------------------------
- Action is invalid, channel update will update action flag too.
- Icon image is invalid, channel update will try to update icon based on feed info.
- So, full update becomes useless. It is removed.

Channel context menu
--------------------
"Delete downloaded file" context menu is only shown when action target of channel is enclosure.

+++ src/free/yhc/feeder/model/Utils.java
@@ -410,13 +410,15 @@ public class Utils {
 
     // return : false(fail to full-delete)
     public static boolean
-    removeFileRecursive(File f) {
+    removeFileRecursive(File f, boolean bDeleteMe) {
         if (f.isDirectory()) {
             for (File c : f.listFiles())
-                if (!removeFileRecursive(c))
+                if (!removeFileRecursive(c, true))
                     return false;
         }
-        return f.delete();
+        if (bDeleteMe)
+            return f.delete();
+        return true;
     }
 
     public static String
